Add files via upload
"This file contains the model's output. Please note that since this model relies on semantic analysis, certain words have been converted to their Hindi spelling (e.g., 'bag' to "बैग"). Consequently, they cannot be reverted to their English spelling. Additionally, due to this semantic approach, some sentences may not undergo further processing. For instance, in the first sentence, the word 'Share' literally translates to 'शेयर करना' when used independently, but in this context, it translates to 'साझा'."